+++ Verizon/src/main/java/com/mopub/mobileads/VerizonAdapterConfiguration.java
@@ -1,5 +1,6 @@
 package com.mopub.mobileads;
 
+import android.app.Application;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -13,15 +14,18 @@ import com.mopub.mobileads.verizon.BuildConfig;
 import com.verizon.ads.Logger;
 import com.verizon.ads.SDKInfo;
 import com.verizon.ads.VASAds;
+import com.verizon.ads.edition.StandardEdition;
+import com.verizon.ads.utils.ThreadUtils;
 
 import java.util.Map;
 
 public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
 
-    public static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
-    public static final String MEDIATOR_ID = "MoPubVAS-" + ADAPTER_VERSION;
-
+    private static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
     private static final String MOPUB_NETWORK_NAME = BuildConfig.NETWORK_NAME;
+    private static final String VAS_SITE_ID_KEY = "siteId";
+
+    static final String MEDIATOR_ID = "MoPubVAS-" + ADAPTER_VERSION;
 
     @NonNull
     @Override
@@ -51,19 +55,17 @@ public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
         }
 
         final String adapterVersion = getAdapterVersion();
-        return (!TextUtils.isEmpty(adapterVersion)) ? adapterVersion.substring(0, adapterVersion.lastIndexOf('.')) : "";
+        return (!TextUtils.isEmpty(adapterVersion)) ? adapterVersion.substring(0,
+                adapterVersion.lastIndexOf('.')) : "";
     }
 
     @Override
-    public void initializeNetwork(@NonNull final Context context, @Nullable final Map<String, String> configuration,
+    public void initializeNetwork(@NonNull final Context context,
+                                  @Nullable final Map<String, String> configuration,
                                   @NonNull final OnNetworkInitializationFinishedListener listener) {
 
+        Preconditions.checkNotNull(context);
         Preconditions.checkNotNull(listener);
-
-        // Due to a limitation in the Verizon Ads SDK with tracking the Activity lifecycle, adapters
-        // will skip initializing the SDK directly.
-        listener.onNetworkInitializationFinished(VerizonAdapterConfiguration.class,
-                MoPubErrorCode.ADAPTER_INITIALIZATION_SUCCESS);
 
         final MoPubLog.LogLevel mopubLogLevel = MoPubLog.getLogLevel();
 
@@ -72,5 +74,33 @@ public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
         } else if (mopubLogLevel == MoPubLog.LogLevel.INFO) {
             VASAds.setLogLevel(Logger.INFO);
         }
+
+        String siteId = null;
+
+        if (configuration != null) {
+            siteId = configuration.get(VAS_SITE_ID_KEY);
+        }
+
+        // The Verizon SDK needs a meaningful siteId to initialize. siteId is cached on the first request.
+        if (TextUtils.isEmpty(siteId)) {
+            listener.onNetworkInitializationFinished(VerizonAdapterConfiguration.class,
+                    MoPubErrorCode.ADAPTER_INITIALIZATION_SUCCESS);
+
+            return;
+        }
+
+        final String finalSiteId = siteId;
+        ThreadUtils.postOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                if (context instanceof Application && StandardEdition.initialize((Application) context, finalSiteId)) {
+                    listener.onNetworkInitializationFinished(VerizonAdapterConfiguration.class,
+                            MoPubErrorCode.ADAPTER_INITIALIZATION_SUCCESS);
+                } else {
+                    listener.onNetworkInitializationFinished(VerizonAdapterConfiguration.class,
+                            MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+                }
+            }
+        });
     }
 }