+++ Verizon/src/main/java/com/mopub/mobileads/VerizonInterstitial.java
@@ -1,6 +1,7 @@
 package com.mopub.mobileads;
 
 import android.app.Activity;
+import android.app.Application;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.text.TextUtils;
@@ -8,6 +9,7 @@ import android.text.TextUtils;
 import com.mopub.common.MoPub;
 import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
+import com.verizon.ads.ActivityStateManager;
 import com.verizon.ads.Bid;
 import com.verizon.ads.BidRequestListener;
 import com.verizon.ads.CreativeInfo;
@@ -69,39 +71,33 @@ public class VerizonInterstitial extends CustomEventInterstitial {
             return;
         }
 
+        // Cache serverExtras so siteId can be used to initalizate VAS early at next launch
+        verizonAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
+
+        String siteId = serverExtras.get(getSiteIdKey());
+        String placementId = serverExtras.get(getPlacementIdKey());
+
         if (!VASAds.isInitialized()) {
-            final String siteId = serverExtras.get(getSiteIdKey());
+            Application application = null;
 
-            if (TextUtils.isEmpty(siteId)) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "Ad request to Verizon failed because " +
-                        "siteId is empty");
-
-                logAndNotifyInterstitialFailed(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR);
-
-                return;
+            if (context instanceof Application) {
+                application = (Application) context;
+            } else if (context instanceof Activity) {
+                application = ((Activity) context).getApplication();
             }
 
-            if (!(context instanceof Activity)) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "Ad request to Verizon failed because " +
-                        "context is not an Activity");
+
+            if (application == null || !StandardEdition.initialize(application, siteId)) {
 
                 logAndNotifyInterstitialFailed(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR);
-
-                return;
-            }
-
-            final boolean success = StandardEdition.initializeWithActivity((Activity) context, siteId);
-
-            if (!success) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "Failed to initialize the Verizon SDK");
-
-                logAndNotifyInterstitialFailed(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR);
-
-                return;
             }
         }
 
-        final String placementId = serverExtras.get(getPlacementIdKey());
+        // The current activity must be set as resumed so VAS can track ad visibility
+        ActivityStateManager activityStateManager = VASAds.getActivityStateManager();
+        if (activityStateManager != null && context instanceof Activity) {
+            activityStateManager.setState((Activity) context, ActivityStateManager.ActivityState.RESUMED);
+        }
 
         if (TextUtils.isEmpty(placementId)) {
             MoPubLog.log(CUSTOM, ADAPTER_NAME, "Ad request to Verizon failed because placement " +
@@ -127,8 +123,6 @@ public class VerizonInterstitial extends CustomEventInterstitial {
         } else {
             interstitialAdFactory.load(bid, new VerizonInterstitialListener());
         }
-
-        verizonAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
     }
 
     /**